+++ pkg/linter/lib/src/rules/unnecessary_async.dart
@@ -193,6 +193,6 @@ class _Visitor extends SimpleAstVisitor<void> {
 extension on InterfaceType {
   bool get isDartAsyncFutureOrSubtype {
     var typeProvider = element.library.typeProvider;
-    return asInstanceOf2(typeProvider.futureElement2) != null;
+    return asInstanceOf2(typeProvider.futureElement) != null;
   }
 }